requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
arrow=1.3.0=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
blinker=1.7.0=pyhd8ed1ab_0
boto3=1.34.88=pyhd8ed1ab_0
botocore=1.34.89=pyge38_1234567_0
bravado=11.0.3=pyhd8ed1ab_0
bravado-core=5.17.1=pyhd8ed1ab_0
brotli=1.1.0=hd590300_1
brotli-bin=1.1.0=hd590300_1
brotli-python=1.0.9=py38h6a678d5_7
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
captum=0.6.0=pyhd8ed1ab_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py38h6d47a40_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=unix_pyh707e725_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.1.1=py38h7f3f72f_1
cryptography=42.0.5=py38hcdda232_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.8.1=py38h17151c0_0
decorator=5.1.1=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h59595ed_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.51.0=py38h01eb140_0
fqdn=1.5.1=pyhd8ed1ab_0
freetype=2.12.1=h4a9f257_0
future=1.0.0=pyhd8ed1ab_0
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.43=pyhd8ed1ab_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=73.2=h59595ed_0
idna=3.4=py38h06a4308_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.1.1=pyhd8ed1ab_0
importlib_metadata=7.1.0=hd8ed1ab_0
importlib_resources=6.1.1=py38h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.12.2=pyh41d4057_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=py38h06a4308_0
jmespath=1.0.1=pyhd8ed1ab_0
joblib=1.2.0=py38h06a4308_0
jpeg=9e=h5eee18b_1
jsonpointer=2.4=py38h578d9bd_3
jsonref=1.1.0=pyhd8ed1ab_0
jsonschema=4.21.1=pyhd8ed1ab_0
jsonschema-specifications=2023.12.1=pyhd8ed1ab_0
jsonschema-with-format=4.21.1=pyhd8ed1ab_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py38h7f3f72f_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.1.0=hd590300_1
libbrotlidec=1.1.0=hd590300_1
libbrotlienc=1.1.0=hd590300_1
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20191231=he28a2e2_2
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=hc881cc4_6
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=13.2.0=hc881cc4_6
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=13.2.0=h95c4c6d_6
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.16=hd590300_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.10.4=hfdd30dd_2
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.3=py38h578d9bd_0
matplotlib-base=3.7.3=py38h58ed7fa_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
monotonic=1.5=py_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
msgpack-python=1.0.3=py38h43d8883_1
munkres=1.1.4=pyh9f0ad1d_0
mysql=5.7.20=hf484d3e_1001
ncurses=6.4=h6a678d5_0
neptune=1.10.2=pyhd8ed1ab_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
nibabel=5.2.1=pyha770c72_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_1
packaging=23.2=py38h06a4308_0
pandas=1.4.2=py38h47df419_1
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.42=hebb0a14_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py38h5eee18b_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.2.0=pyhd8ed1ab_0
pooch=1.7.0=py38h06a4308_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
psutil=5.9.8=py38h01eb140_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_1
pyparsing=3.1.2=pyhd8ed1ab_0
pyqt=5.15.4=py38hfa26641_0
pyqt5-sip=12.9.0=py38hfa26641_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.2.2=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py38h5eee18b_0
pyzmq=26.0.2=py38ha44f8e3_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
referencing=0.34.0=pyhd8ed1ab_0
requests=2.31.0=py38h06a4308_1
requests-oauthlib=2.0.0=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3987=1.3.8=py_0
rpds-py=0.18.0=py38h0cc4f7c_0
s3transfer=0.10.1=pyhd8ed1ab_0
scikit-learn=1.3.0=py38h1128e8f_1
scipy=1.10.1=py38hf6e8229_1
setuptools=68.2.2=py38h06a4308_0
simplejson=3.19.2=py38h01eb140_0
sip=6.5.1=py38h709712a_2
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pyhd8ed1ab_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
swagger-spec-validator=3.0.3=pyhd8ed1ab_0
sympy=1.12=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=2.2.2=py38_cu118
torchtriton=2.2.0=py38
torchvision=0.17.2=py38_cu118
tornado=6.4=py38h01eb140_0
traitlets=5.14.3=pyhd8ed1ab_0
types-python-dateutil=2.9.0.20240316=pyhd8ed1ab_0
typing=3.10.0.0=pyhd8ed1ab_0
typing-extensions=4.9.0=py38h06a4308_1
typing_extensions=4.9.0=py38h06a4308_1
unicodedata2=15.1.0=py38h01eb140_0
uri-template=1.3.0=pyhd8ed1ab_0
urllib3=1.26.18=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pyhd8ed1ab_0
websocket-client=1.7.0=pyhd8ed1ab_0
wheel=0.41.2=py38h06a4308_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h59595ed_1
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
